requirements.txt
customtkinter
rembg
Pillow
onnxruntime
